requirements.txt
beautifulsoup4
requests
selenium
python-dotenv
pytest-playwright